+++ lab9kos/Models/Domain/ITaakRepository.cs
@@ -8,7 +8,7 @@ namespace lab9kos.Models.Domain
     public interface ITaakRepository
     {
         void AddTaak(Taak taak);
-        Taak GetBy(int id);
+        Taak GetBy(long id);
         void RemoveTaak(Taak taak);
         void RemoveTaakGebruiker(TaakGebruiker taakGebruiker);
         List<Taak> GetAllWithNiveau(TaakRealisatieNiveau niveau);
